refactor(customer): reorganize event files into domain-specific directories
Move customer domain event files into their respective domain directories (customer, wishlist, review) for better organization
Remove duplicate handler registrations in customer module
Update imports to use new file paths
