+++ src/domains/cart/aggregates/events/index.ts
@@ -1,5 +1,5 @@
 export { CartCreatedEvent } from './cart-created.event';
-export { CustomerCreatedEvent } from '../../../customer/aggregates/events/customer-created.event';
+export { CustomerCreatedEvent } from 'src/domains/customer/aggregates/events';
 export { ItemRemovedFromCartEvent } from './remove-item-cart.event';
 export { AddItemToCartEvent } from './add-item-to-cart.event';
 export { ItemQuantityUpdatedEvent } from './item-quantity-updated.event';
