+++ src/domains/customer/aggregates/events/customer/customer-updated.event.ts
@@ -1,4 +1,4 @@
-import { Customer } from '../entities';
+import { Customer } from '../../entities';
 
 export class CustomerUpdatedEvent {
   constructor(public readonly customer: Customer) {}
