+++ src/domains/customer/aggregates/events/index.ts
@@ -1,8 +1,8 @@
-export { CustomerCreatedEvent } from './customer-created.event';
-export { CustomerUpdatedEvent } from './customer-updated.event';
-export { WishlistItemCreatedEvent } from './wishlist-created.event';
-export { WishlistItemDeletedEvent } from './wishlist-deleted.event';
-export { WishlistManyItemsDeletedEvent } from './wishlist-many-items-deleted.event';
-export { CustomerReviewProductCreatedEvent } from './review-created.event';
-export { CustomerReviewProductUpdatedEvent } from './review-updated.event';
-export { CustomerReviewProductDeletedEvent } from './review-deleted.event';
+export { CustomerCreatedEvent } from './customer/customer-created.event';
+export { CustomerUpdatedEvent } from './customer//customer-updated.event';
+export { WishlistItemCreatedEvent } from './wishlist/wishlist-created.event';
+export { WishlistItemDeletedEvent } from './wishlist/wishlist-deleted.event';
+export { WishlistManyItemsDeletedEvent } from './wishlist/wishlist-many-items-deleted.event';
+export { CustomerReviewProductCreatedEvent } from './review/review-created.event';
+export { CustomerReviewProductUpdatedEvent } from './review/review-updated.event';
+export { CustomerReviewProductDeletedEvent } from './review/review-deleted.event';
