+++ src/domains/customer/aggregates/events/review/review-updated.event.ts
@@ -1,5 +1,5 @@
-import { Customer } from '../entities';
-import { CustomerReviewProduct } from '../value-objects/customer-review-product.vo';
+import { Customer } from '../../entities';
+import { CustomerReviewProduct } from '../../value-objects/customer-review-product.vo';
 
 export class CustomerReviewProductUpdatedEvent {
   constructor(
